Setup ActionBars in layoutlib the same way the platform does it. do not merge.
Instead of using a simple ImageView for the icon, this uses the platform
layout/action_bar_home which uses a custom class to position and resize
the icon (and also supports the Up icon that we don't yet support).

This ensures that the icon is properly positionned and sized like
on devices.

(cherry picked from commit 2d4da5404d8983444a4cd2cb8d6df1021a50c32b)

Change-Id: Ifd3bc318089b70ba843519523e366e59d434e919