Fix deadlock in PDU code
Bug 6257442

If PduPersister.load bails out early and throws an exception,
it will leave the uri it's working on in an updated-locked
state. With this change, the uri is removed from the updated
list in a finally clause. The other PDU function in
this file already does the correct thing.

Change-Id: I596bd46724282b245ddba109670f3cebd61ed57c