After reboot with airplane mode and BT enabled, keep BT enabled.
Enabling airplane mode disables Wi-Fi and Bluetooth as well as the
cell radio. Bluetooth and Wi-Fi can then be reenabled if desired.
After a reboot, Bluetooth was always disabled if airplane mode
was enabled. Changed to set Bluetooth to its previous state after
reboot, regardless of the airplane mode setting. This is already
the behavior for Wi-Fi after a reboot.

Bug: 5623624
Change-Id: I1c44500221aa0e5d1d7a5c9046d6c1380d28b5c0